Fixed ipset updating
Previously ipset updates to ipsets that already existed tried to destroy and then create the list.  You cannot destroy a list actively in use by IPTables.  In this case we checked the diff between the active ipset and the proposed ipset, if it found a difference it now creates a new temporary in-memory ipset and then swaps them in-place and then proceeds to remove that temporary in-memory ipset.