If possible, swap set contents instead of destroying a set and then recreating it
You can not destroy a set in use (eg. by iptables). Also, this disrupts
potentially the usage (since an empty set is, well, empty). The correct
approach is to create a new set, fill it with the correct content, and
then use `ipset swap from-set to-set`.

This is only possible if the set's hash structure did not change... If
the new and old one have different structures, you have other problems
to solve too.